+++ MarketOrderTreeModel.h
@@ -103,7 +103,7 @@ namespace Evernus
 
         QString getGroupName(EveType::IdType typeId) const;
 
-        static uint getStationId(const MarketOrder &order);
+        static quint64 getStationId(const MarketOrder &order);
 
     private:
         Character::IdType mCharacterId = Character::invalidId;
